requirements.txt
fastapi>=0.115.0
uvicorn>=0.34.0
typer>=0.12.0
pyyaml>=6.0.0
